강화 정보 초기화 로직

+++ Assets/신원석/Scripts/Config/FoodData.cs
@@ -23,7 +23,7 @@ public class EnhanceStepData
 {
     public string indate;
     public int step;
-    public int successRate;
+    public float successRate;
     public List<EnhanceMaterialData> ingredients;
     public int cost;
 }
